+++ src/fleet/azext_fleet/_params.py
@@ -13,7 +13,6 @@ from azure.cli.core.commands.parameters import (
 from azure.cli.core.commands.validators import get_default_location_from_resource_group
 
 from azext_fleet._validators import validate_member_cluster_id, validate_upgrade_type, validate_kubernetes_version
-from azext_fleet._completers import get_k8s_versions_completion_list
 
 
 def load_arguments(self, _):
@@ -50,7 +49,7 @@ def load_arguments(self, _):
 
     with self.argument_context('fleet updaterun create') as c:
         c.argument('upgrade_type', validator=validate_upgrade_type)
-        c.argument('kubernetes_version', validator=validate_kubernetes_version, completer=get_k8s_versions_completion_list)
+        c.argument('kubernetes_version', validator=validate_kubernetes_version)
         c.argument('stages', type=file_type, completer=FilesCompleter())
         c.argument('update_group')
 
@@ -66,5 +65,5 @@ def load_arguments(self, _):
 
     with self.argument_context('fleet updaterun create') as c:
         c.argument('upgrade_type', validator=validate_upgrade_type)
-        c.argument('kubernetes_version', validator=validate_kubernetes_version, completer=get_k8s_versions_completion_list)
+        c.argument('kubernetes_version', validator=validate_kubernetes_version)
         c.argument('stages', type=file_type, completer=FilesCompleter())